api-server/requirements.txt
Flask==2.1.2
Flask-Cors==3.0.10
flask-restx==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.2.3
Flask-Migrate==3.1.0
Flask-Limiter==2.4.0
Werkzeug==2.1.2
SQLAlchemy==1.4.22
PyJWT==2.1.0
python-dotenv==1.1.0
bleach==6.2.0